Renamed WindowSDL2 to SDL2Window
Closes https://github.com/PistonDevelopers/sdl2_window/issues/77
